Speak with csv file/requirements.txt
pandas
streamlit
openai
python-dotenv
langchain
langchain-openai
langchain-experimental
